requirements.txt
Flask
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
oauth2client
python-dotenv